event_service/requirements.txt
redis
sqlalchemy
asyncpg
python-dotenv
pydantic
httpx


